converted class component into functional component and added separate files for setting
